requirements.txt
beautifulsoup4==4.12.2
Django==5.0
django-autoslug==1.9.9
django-health-check==3.17.0
gunicorn==21.2.0
html5lib==1.1
lxml==4.9.4
psycopg2-binary==2.9.9
requests==2.31.0
tenacity==8.2.3
whitenoise==6.6.0
